Refactor: extracted code from we_term.c into we_screen.c plus minor renames of functions. * Translated col_kasten to colorsets. It is an educated guess. * Transformed some macro's into functions. (term_move and term_refresh). * Minor typo corrections.
