+++ src/we_unix.h
@@ -34,7 +34,7 @@ char* e_file_info(char* filen, char* str, int* num, int sw);
 void ini_repaint(we_control_t* control);
 void end_repaint(void);
 int e_frb_t_menue(int sw, int xa, int ya, we_window_t* window, int md);
-void e_pr_t_col_kasten(int xa, int ya, int x, int y, we_window_t* window, int sw);
+void e_pr_t_colorsets(int xa, int ya, int x, int y, we_window_t* window, int sw);
 int e_ini_unix(int* argc, char** argv);
 int e_recover(we_control_t* control);
 int e_ini_schirm(int argc, char** argv);
@@ -46,7 +46,7 @@ extern int (*u_bioskey)(void);
 extern int (*e_frb_u_menue)(int sw, int xa, int ya, we_window_t* window, int md);
 extern we_color_t (*e_s_u_clr)(int fg_color, int bg_color);
 extern we_color_t (*e_n_u_clr)(int fg_bg_color);
-extern void (*e_pr_u_col_kasten)(int xa, int ya, int x,
+extern void (*e_pr_u_colorsets)(int xa, int ya, int x,
                                  int y, we_window_t* window, int sw);
 extern int (*fk_mouse)(int g[]);
 extern int (*e_u_refresh)(void);
